Доделал контроллеры (кроме пользователей ибо там надо много че перерабатывать), привел все ошибки к одному виду

+++ CRM.WebAPI/CRM.WebAPI/Controllers/ContactController.cs
@@ -1,12 +1,9 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using CRM.Domain.Models;
+using CRM.Domain.ModelsToUpload;
 using CRM.Infrastructure.Interfaces;
 using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using ConflictResult = CRM.Core.Implement.ConflictResult;
 
 namespace CRM.WebAPI.Controllers
 {
@@ -34,48 +31,48 @@ namespace CRM.WebAPI.Controllers
         public async Task<ActionResult> GetById(Guid id)
         {
             if (id == Guid.Empty)
-                return BadRequest("Id is empty");
+                return BadRequest(new ConflictResult("Empty input is not allowed!"));
             var result = await _repository.GetById(id);
             if (result.Successful)
                 return Ok(result);
 
-            return NotFound($"Contact with id {id} does not exist");
+            return NotFound(result);
         }
 
         // PUT api/Contact/5
         [HttpPut]
-        public async Task<ActionResult> Insert([FromBody] Contact contact)
+        public async Task<ActionResult> Insert([FromBody] ContactModel contact)
 
         {
-
+            if (contact.Name == "" || contact.Lastname == "")
+                return BadRequest(new ConflictResult("Empty input is not allowed!"));
             var result = await _repository.Put(contact);
             if (result.Successful)
                 return Ok(result);
-            return BadRequest(result.ErrorMessage);
+            return BadRequest(result);
         }
 
         [HttpPost]
         public async Task<ActionResult> Update([FromBody] Contact dataToUpdate)
         {
-            var result =  await _repository.Update(dataToUpdate);
-            if (!result.Successful)
-            {
-                return BadRequest(result.ErrorMessage);
-            }
-            return Ok();
+            if (dataToUpdate.ContactId == Guid.Empty)
+                return BadRequest(new ConflictResult("Empty input is not allowed!"));
+            var result = await _repository.Update(dataToUpdate);
+            if (result.Successful)
+                return Ok(result);
+            return BadRequest(result);
         }
 
         // DELETE api/Contact/5
         [HttpDelete("{id}")]
         public async Task<ActionResult> Delete(Guid id)
         {
-            var result =  await _repository.RemoveById(id);
-            if (!result.Successful)
-            {
-                return BadRequest(result.ErrorMessage);
-            }
-           
-            return Ok();
+            if (id == Guid.Empty)
+                return BadRequest(new ConflictResult("Empty input is not allowed!"));
+            var result = await _repository.RemoveById(id);
+            if (result.Successful)
+                return Ok(result);
+            return BadRequest(result);
         }
     }
 }
