+++ CRM.WebAPI/CRM.WebAPI/Controllers/LoginController.cs
@@ -3,10 +3,12 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using CRM.Domain.Models;
+using CRM.Domain.ModelsToUpload;
 using CRM.Infrastructure.Interfaces;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using ConflictResult = CRM.Core.Implement.ConflictResult;
 
 
 namespace CRM.WebAPI.Controllers
@@ -35,22 +37,25 @@ namespace CRM.WebAPI.Controllers
         public async Task<ActionResult> GetById(Guid id)
         {
             if (id == Guid.Empty)
-                return BadRequest("Id is empty");
+                return BadRequest(new ConflictResult("Empty input is not allowed!"));
             var result = await _repository.GetById(id);
             if (result.Successful)
                 return Ok(result);
-            return NotFound(result.ErrorMessage);
+
+            return NotFound(result);
         }
 
 
         // PUT: api/Login/5
         [HttpPut]
-        public async Task<ActionResult> Insert([FromBody] Login login)
+        public async Task<ActionResult> Insert([FromBody] LoginModel login)
         {
+            if(login.UserId == Guid.Empty || login.LoginTime == null)
+                return BadRequest(new ConflictResult("Empty input is not allowed!"));
             var result = await _repository.Put(login);
             if (result.Successful)
                 return Ok(result);
-            return BadRequest(result.ErrorMessage);
+            return BadRequest(result);
         }
 
         // POST: api/Login
@@ -58,20 +63,24 @@ namespace CRM.WebAPI.Controllers
         public async Task<ActionResult>
             Update([FromBody] Login dataToUpdate)
         {
+            if (dataToUpdate.LoginId == Guid.Empty)
+                return BadRequest(new ConflictResult("Empty input is not allowed!"));
             var result = await _repository.Update(dataToUpdate);
             if (result.Successful)
                 return Ok(result);
-            return BadRequest(result.ErrorMessage);
+            return BadRequest(result);
         }
 
         // DELETE: api/Login/5
         [HttpDelete("{id}")]
         public async Task<ActionResult> Delete(Guid id) //удаляет по id
         {
+            if (id == Guid.Empty)
+                return BadRequest(new ConflictResult("Empty input is not allowed!"));
             var result = await _repository.RemoveById(id);
             if (result.Successful)
                 return Ok(result);
-            return BadRequest(result.ErrorMessage);
+            return BadRequest(result);
         }
     }
 }
