+++ CRM.WebAPI/CRM.WebAPI/Controllers/PsychologistController.cs
@@ -3,13 +3,16 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using CRM.Domain.Models;
+using CRM.Domain.ModelsToUpload;
 using CRM.Infrastructure.Interfaces;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using ConflictResult = CRM.Core.Implement.ConflictResult;
 
 namespace CRM.WebAPI.Controllers
-{[Route("api/[controller]/[action]")]
+{
+    [Route("api/[controller]/[action]")]
     [ApiController]
     [Authorize]
     public class PsychologistController : ControllerBase
@@ -20,6 +23,7 @@ namespace CRM.WebAPI.Controllers
         {
             _repository = repository ?? throw new ArgumentException(nameof(repository));
         }
+
         // GET: api/Psychologist
         [HttpGet]
         public async Task<ActionResult> GetAll()
@@ -32,43 +36,51 @@ namespace CRM.WebAPI.Controllers
         public async Task<ActionResult> GetById(Guid id)
         {
             if (id == Guid.Empty)
-                return BadRequest("Id is empty");
+                return BadRequest(new ConflictResult("Empty input is not allowed!"));
             var result = await _repository.GetById(id);
             if (result.Successful)
                 return Ok(result);
-            return NotFound(result.ErrorMessage);
+
+            return NotFound(result);
         }
-        
+
 
         // PUT: api/Psychologist/5
         [HttpPut]
-        public async Task<ActionResult>Insert([FromBody]Psychologist psychologist)
+        public async Task<ActionResult> Insert([FromBody] PsychologistModel psychologist)
         {
+            if (psychologist.ContactId == Guid.Empty || psychologist.ContactId == Guid.Empty ||
+                psychologist.UserId == Guid.Empty)
+                return BadRequest(new ConflictResult("Empty input is not allowed!"));
             var result = await _repository.Put(psychologist);
             if (result.Successful)
                 return Ok(result);
-            return BadRequest(result.ErrorMessage);
+            return BadRequest(result);
         }
+
         // POST: api/Psychologist
         [HttpPost]
         public async Task<ActionResult>
-            Update([FromBody]Psychologist dataToUpdate)
+            Update([FromBody] Psychologist dataToUpdate)
         {
+            if (dataToUpdate.PsychologistId == Guid.Empty)
+                return BadRequest(new ConflictResult("Empty input is not allowed!"));
             var result = await _repository.Update(dataToUpdate);
             if (result.Successful)
                 return Ok(result);
-            return BadRequest(result.ErrorMessage);
-
+            return BadRequest(result);
         }
 
         // DELETE: api/Psychologist/5
         [HttpDelete("{id}")]
-        public async Task<ActionResult> Delete(Guid id) 
+        public async Task<ActionResult> Delete(Guid id)
         {
+            if (id == Guid.Empty)
+                return BadRequest(new ConflictResult("Empty input is not allowed!"));
             var result = await _repository.RemoveById(id);
             if (result.Successful)
                 return Ok(result);
-            return BadRequest(result.ErrorMessage);
+            return BadRequest(result);
         }
     }
 }
