+++ CRM.WebAPI/CRM.WebAPI/Controllers/VisitController.cs
@@ -1,8 +1,9 @@
 ﻿using CRM.Domain.Models;
+using CRM.Domain.ModelsToUpload;
 using CRM.Infrastructure.Interfaces;
-using CRM.WebAPI.ModelsToUpload;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using ConflictResult = CRM.Core.Implement.ConflictResult;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
@@ -14,49 +15,13 @@ namespace CRM.WebAPI.Controllers
     public class VisitController : ControllerBase
     {
         private IVisitRepository _repository;
-   
-       
+
         public VisitController(IVisitRepository repository)
         {
             _repository = repository ?? throw new ArgumentException(nameof(repository));
         }
-        
-        //Logic
-        // [HttpPost]
-        // public async Task<ActionResult> CreateVisit([FromForm] VisitModel visit)
-        // {
-        //     try
-        //     {
-        //
-        //        await _repository.Put(new Visit()
-        //         {
-        //             VisitId = new Guid(),
-        //             ClientId = visit.ClientId,
-        //             ClientNote = visit.ClientNote,
-        //             DateTime = visit.DateTime,
-        //             PsychologistDescription = visit.Description,
-        //             PsychologistId = visit.PsychologistId,
-        //             ServiceId = visit.ServiceId
-        //         });
-        //         return Ok();
-        //     }
-        //     catch (Exception e)
-        //     {
-        //         Console.WriteLine(e);
-        //         throw;
-        //     }
-        // }
-        
-        
-        
-        
-        
-        
-        
-        
-        
-        //Cruds
-        
+
+
         // GET: api/Visit
         [HttpGet]
         public async Task<ActionResult> GetAll()
@@ -69,42 +34,50 @@ namespace CRM.WebAPI.Controllers
         public async Task<ActionResult> GetById(Guid id)
         {
             if (id == Guid.Empty)
-                return BadRequest("Id is empty");
+                return BadRequest(new ConflictResult("Empty input is not allowed!"));
             var result = await _repository.GetById(id);
             if (result.Successful)
                 return Ok(result);
-            return NotFound(result.ErrorMessage);
+
+            return NotFound(result);
         }
 
 
         // PUT api/Visit/5
         [HttpPut]
-        public async Task<ActionResult> Insert([FromBody] Visit visit)
+        public async Task<ActionResult> Insert([FromBody] VisitModel visit)
 
         {
+            if (visit.ClientId == Guid.Empty || visit.PsychologistId == Guid.Empty || visit.ServiceId == Guid.Empty ||
+                visit.DateTime == null)
+                return BadRequest(new ConflictResult("Empty input is not allowed!"));
             var result = await _repository.Put(visit);
             if (result.Successful)
                 return Ok(result);
-            return BadRequest(result.ErrorMessage);
+            return BadRequest(result);
         }
 
         [HttpPost]
         public async Task<ActionResult> Update([FromBody] Visit dataToUpdate)
         {
+            if (dataToUpdate.VisitId == Guid.Empty)
+                return BadRequest(new ConflictResult("Empty input is not allowed!"));
             var result = await _repository.Update(dataToUpdate);
             if (result.Successful)
                 return Ok(result);
-            return BadRequest(result.ErrorMessage);
+            return BadRequest(result);
         }
 
         // DELETE api/Visit/5
         [HttpDelete("{id}")]
         public async Task<ActionResult> Delete(Guid id)
         {
+            if (id == Guid.Empty)
+                return BadRequest(new ConflictResult("Empty input is not allowed!"));
             var result = await _repository.RemoveById(id);
             if (result.Successful)
                 return Ok(result);
-            return BadRequest(result.ErrorMessage);
+            return BadRequest(result);
         }
     }
 }
